+++ package.json
@@ -5,21 +5,13 @@
   "main": "server.js",
   "dependencies": {
     "express": "^4.16.3",
-    "mongoose": "^5.3.4"
+    "mongoose": "^5.3.0"
   },
   "devDependencies": {},
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node server.js"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/DishantaK/gdoc_clone.git"
-  },
-  "author": "Dishanta K , Matt B , Christian G, Marcus R",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/DishantaK/gdoc_clone/issues"
-  },
-  "homepage": "https://github.com/DishantaK/gdoc_clone#readme"
+  "author": "mandeep kaur",
+  "license": "ISC"
 }
