+++ public/css/doc.css
@@ -60,20 +60,18 @@
 
 #options  {
     display: grid;
-    grid-template-columns: repeat(10, 1fr) ;
-    /* width: 80%; */
+    grid-template-columns: repeat(9, 1fr);
+    width: 55%;
     list-style-type: none;
     margin-left: -40px;
     justify-content: left;
-
 }
 
 #options button  {
-    background-color: #fff;
+    /* background-color: #fff;
     padding: 0px;
     border:none;
-    margin: 0px;
-
+    margin: 0px; */
 }
 
 #options button:hover {
