--- public/css/newdoc.css
@@ -1,163 +1,0 @@
-#docSection {
-    background-color: #f4f4f4;
-    padding: 0px;
-  
-}
-
-#docHeader {
-    background-color: #fff;
-    height: 120px;
-    padding: 0px;
-    border-bottom: 1px solid rgb(221, 221, 221);
-    
-}
-
-#measure {
-    height: 20px;
-    border-bottom: 1px solid rgb(221, 221, 221);
-    margin-bottom: 15px;
-    display: grid;
-    width: 100%;
-    
-}
-
-#docArea {
-    padding: 30px;
-    width: 90%;
-    background-color: #fff;
-    margin:auto;
-    box-shadow: 0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24);
-    height: 95vh;
-}
-
-#docAreaText {
-    background-color: #fff;
-    width: 100%;
-    border: none;
-}
-
-#docTitle {
-    height:70%;
-    display: grid;
-    grid-template-columns: 7% 1fr;
-    grid-template-rows: repeat(2,1fr);
-    border-bottom: 1px solid rgb(221, 221, 221);
-   
-}
-#docTitle a {
-    grid-row: span 2;
-    grid-template-columns: span 1;
-}
-#docTitle img {
-    width: 50px;
-    height: 50px;
-    margin: 20px;
-}
-
-#docTitle input {
-    background-color: transparent;
-    border: none;
-    width: 25%;
-    grid-column: span 1;
-}
-
-#docOpt {
-    height:30%;
-}
-
-#options  {
-    grid-column: span 1;
-    grid-row: span 1;
-    display: grid;
-    grid-gap: 10px;
-    width: 70%;
-    grid-template-columns: repeat(9, 1fr) 10%;
-    list-style-type: none;
-    /* justify-content: left; */
-
-
-}
-
-#options li {
-    white-space: nowrap;
-}
-#options li:hover {
-    /* background-color:  rgb(221, 221, 221); */
-}
-
-
-
-#options button  {
-    background-color: #fff;
-    padding: 0px;
-    border:none;
-    margin: 0px;
-    color:#000;
-
-}
-
-#options button:hover {
-    cursor: pointer;
-}
-.caret {
-    display: none;
-}
-
-
-
-
-@media only screen  and (max-width: 500px){
-    #docHeader {
-        padding: 2px;
-        margin-bottom: 25px;
-        
-    }
-   #docHeader input {
-       align-self: center;
-       width: 50%;
-       margin-left: 60px;
-       margin-bottom: 5px;
-   }
-   #docTitle {
-    margin-bottom: 15px;
-    border-bottom: none;
-   }
-    #styleOpt{
-        display: grid;
-        grid-template-columns: 1fr;
-        margin-top: 5px;
-    }
-
-    #styleOpt select {
-        width: 100%;
-    }
-
-    #options {
-        grid-template-columns: repeat(3,1fr);
-        justify-content: right;
-        margin-bottom: 10px;
-        margin-left: -15px;
-        justify-self: center;
-    }
-
-    #options li {
-        
-        display: block;
-        
-        
-    }
-    
-    #options li:nth-child(4) {
-        display: none;
-    }
-    #options li:nth-child(5) {
-        display: none;
-    }
-    #options li:nth-child(6) {
-        display: none;
-    }
-    #options li:nth-child(7) {
-        display: none;
-    }
-    
-}
