+++ public/css/style.css
@@ -1,3 +1,17 @@
+.doc-btn {
+    position: relative;
+    float: right;
+    background: blue;
+    padding: 10px;
+    margin-right: 20px;
+    color: #fff;
+}
+
+.doc-btn:hover {
+    cursor: pointer;
+    background: green;
+}
+
 .modal-fullpage {
     z-index: 999;
     width: 100%;
@@ -50,8 +64,17 @@
 
 .open-doc-btn {
     display: inline-block;
-    border: 1px solid #000;
+    border: 1px solid #eee;
+    background: #fff;
     cursor: pointer;
+    overflow: hidden;
+    padding: 10px;
+    margin: 10px 20px 10px 0px;
+}
+
+.open-doc-btn:hover {
+    border: 1px solid #7092F1;
+    text-decoration: none;
 }
 
 body {
@@ -81,8 +104,6 @@ h3 {
     font-weight: bold;
     margin-top: 10px;
 }
-
-/* Main Header and Search area */
 
 header {
     height: 80px;
@@ -147,7 +168,6 @@ header {
     background-color: #3C3C3C;
     padding: 25px 70px;
     color: rgb(212, 212, 212);
-    
 }
 
 .new h2 {
@@ -172,14 +192,6 @@ header {
     display: grid;
     grid-template-columns: repeat(4,1fr);
     align-items: center;
-}
-
-#newInner a {
-    text-decoration: none;
-    color: #fff;
-}
-#newInner a:active {
-    text-decoration: none;
 }
 
 #newBlank {
@@ -235,21 +247,16 @@ main  {
     grid-gap: 25px; */
 }
 
-/*  Card  */
-
 .docPrev {
     width: 200px;
     height: 260px;
     background-color: #fff;
-    display: grid;
-    /* grid-template-columns: 1fr;
-    grid-template-rows: 70% 30%; */
-    box-shadow: 0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24);
     color: #3C3C3C;
+    overflow: hidden;
+    font-size: 6px;
 }
 
 .docPrev:hover {
-    border: 1px solid #7092F1;
     cursor: pointer;
 }
 
@@ -263,11 +270,8 @@ h4 {
 }
 .captionLabel {
     border-top: 1px solid rgb(221, 221, 221);
-    grid-row:span 1;
-    display: grid;
-    grid-template-columns: 1fr;
-    grid-template-rows: repeat(2,1fr);
     padding: 15px;
+    margin-top: -100px;
 }
 
 .labelGroup {
@@ -315,8 +319,8 @@ h4 {
         padding-left: 70px;
         background-color: rgb(248, 248, 248);
         font-size: 1em;
-        
     }
+
     .search {
         align-self: start;
         grid-column: span 1;
@@ -334,19 +338,19 @@ h4 {
     .sort {
         position: relative;
     }
+
     #downSort {
         display: inline;
         justify-self: start;
         align-self: start;
         position: absolute;
-        
     }
+
     #existingDocs {
         display: grid;
         grid-template-columns: repeat(4,1fr);
         width: 80%;
-        
-        
     }
+    
 }
