+++ public/doc.html
@@ -9,24 +9,25 @@
     crossorigin="anonymous">
   <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/all.css" integrity="sha384-mzrmE5qonljUremFsqc01SB46JvROS7bZs3IO2EmfFsd15uHvIt+Y8vEf7N7fWAU"
     crossorigin="anonymous">
+    <link href="https://fonts.googleapis.com/css?family=Charmonman|Kodchasan|K2D" rel="stylesheet">
   <link href="https://fonts.googleapis.com/css?family=Noto+Sans+JP:100,400" rel="stylesheet">
   <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
-  <link rel="stylesheet" href="css/style.css" />
-  <link rel="stylesheet" href="css/newdoc.css" />
+  <link rel="stylesheet" href="/css/style.css" />
+  <link rel="stylesheet" href="/css/doc.css" />
   <title>Edit: Document</title>
 </head>
 
 <body>
-  <section id="gdocEdit">
+
+  <form id="gdocEdit">
     <!-- This is where the document editor loads -->
-  </section>
+  </form>
 
   <script src="https://code.jquery.com/jquery-3.3.1.min.js" integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8="
     crossorigin="anonymous"></script>
-  <script src="/js/test.js"></script>  
-  <script src="/js/doc.js"></script>
-  <!-- <script src="/js/app.js"></script> -->
+  <script src="/js/document.js"></script>
 
 </body>
 
 </html>
+
