--- public/indexRES.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <link rel="stylesheet" href="reset.css">
-  <link rel="stylesheet" href="style.css">
-  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/all.css" integrity="sha384-mzrmE5qonljUremFsqc01SB46JvROS7bZs3IO2EmfFsd15uHvIt+Y8vEf7N7fWAU"
-    crossorigin="anonymous">
-  <title>Clone: Google Docs</title>
-</head>
-
-<body>
-
-  <form id="docForm">
-    <div class="title-header">
-      <input type="text" placeholder="Untitled document" name="Document Title" id="input-title">
-    </div>
-    <div class="page">
-      <input type="textarea" placeholder="Document Content" name="Document Content" id="input-content" />
-      <button type="submit" id="add-button"><i class="fas fa-share"></i></button>
-    </div>
-  </form>
-
-  <div class="side-right">
-    <ul id="gdoclist">
-    </ul>
-  </div>
-
-  <script src="https://code.jquery.com/jquery-3.3.1.js" integrity="sha256-2Kok7MbOyxpgUVvAk/HJ2jigOSYS2auK4Pfzbm7uH60="
-    crossorigin="anonymous"></script>
-  <script src="js/app.js"></script>
-
-</body>
-
-</html>
