--- public/newdoc.html
@@ -1,93 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO"
-    crossorigin="anonymous">
-    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/all.css" integrity="sha384-mzrmE5qonljUremFsqc01SB46JvROS7bZs3IO2EmfFsd15uHvIt+Y8vEf7N7fWAU"
-    crossorigin="anonymous">
-    <link href="https://fonts.googleapis.com/css?family=Noto+Sans+JP:100,400" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
-    <link href="https://fonts.googleapis.com/css?family=Charmonman|Kodchasan|K2D" rel="stylesheet">
-    <link rel="stylesheet" href="css/newdoc.css" />
-    <title>Document</title>
-</head>
-<body>
-  
-        <header id="docHeader">
-                <section id="docTitle">
-
-                 <a href="/"><img src="images/docs_48dp.png" placeholder="docs" /></a>
-                 <input type="text" placeholder="Untitled document" />
-
-                 <ul id="options">
-                     <li><div class="btn-group">
-                        <button class="btn btn-secondary btn-sm dropdown-toggle" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                          File
-                        </button>
-                        <div class="dropdown-menu">
-                            <a class="dropdown-item" href="#">Save</a>
-                        </div>
-                      </div></li>
-                     <li><button class="mainOption">Edit</button></li>
-                     <li><button class="mainOption">View</button></li>
-                     <li><button class="mainOption">Insert</button></li>
-                     <li><button class="mainOption">Format</button></li>
-                     <li><button class="mainOption">Tools</button></li>
-                     <li><button class="mainOption">Table</button></li>
-                     <li><button class="mainOption">Add-Ons</button></li>
-                     <li><button class="mainOption">Help</button></li>
-                     <li><p id="status" contenteditable="false"></p></li>
-                   </ul>
-                </section>
-                <section id="styleOpt">
-                <span>
-                    <select id='thefonts' name=fontdown>
-                        <option value="K2D">Select Font</option>
-                        <option value="Times New Roman">Times New Roman</option>
-                        <option value="Courier">Courier</option>
-                        <option value="Georgia">Georgia</option>
-                        <option value="Charmonman">Charmonman</option>
-                        <option value="Kodchasan">Kodchasan</option>
-                    </select>
-                </span>
-                <span>
-                    <select id='fontcolor' name=colordown>
-                        <option value="black">Select Font Color</option>
-                        <option value="blue">Blue</option>
-                        <option value="yellow">Yellow</option>
-                        <option value="red">Red</option>
-                        <option value="orange">Orange</option>
-                        <option value="green">Green</option>
-                    </select>
-                </span>
-                <span>
-                    <select id='fontsize' name=sizedown>
-                        <option value= 3>Select Font Size</option>
-                        <option value= 2>x.5</option>
-                        <option value= 4>x2</option>
-                        <option value= 5>x3</option>
-                        <option value= 6>x4</option>
-                        <option value= 7>x5</option>
-                        </select>
-                </span>
-                </section>
-           </header>
-
-           <main id="docSection">  
-            <div id="measure"></div>
-            <form>
-            <div id="docArea" contenteditable="true"></div>
-            <input type="hidden" id="docAreaText" />
-            </form>
-         </main>
-
-             <script src="https://code.jquery.com/jquery-3.3.1.min.js" integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8=" crossorigin="anonymous"></script>      
-             <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
-             <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
-            <script src= "js/newdoc.js" > </script>
-    
-</body>
-</html>
