+++ routes/html-routes.js
@@ -1,24 +1,15 @@
-//Require path for index.html
+//Require 'path' for index.html
 const path = require('path');
 
-
-/**
- * La middleware.
- * @module exports  HTML routes
- */
+//Pass in 'app' for Express
 module.exports = function (app) {
-    
-       /**
-     * Get Home and Doc Pages.
-     * @async get - gets page // sends html.
-     */
-
-    app.get('/doc/:id', function (req, res) {
+   
+    app.get('/doc', function (req, res) {
         res.sendFile(path.join(__dirname, '../public/doc.html'));
     });
 
-    app.get('/theNewDoc/', function (req, res){
-        res.sendFile(path.join(__dirname, '../public/newdoc.html'));
+    app.get('/doc/:id', function (req, res) {
+        res.sendFile(path.join(__dirname, '../public/doc.html'));
     });
 
      //Default to index.html
@@ -27,3 +18,4 @@ module.exports = function (app) {
     });
 
 }; 
+
