--- routes/routes.js
@@ -1,9 +1,0 @@
-const path = require('path');
-
-module.exports = function(app) {
-   
-    app.get('/', function(req, res) {
-        res.sendFile(path.join(__dirname, '../public/index.html'));
-    });
-
-};
